% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{aggregations}
\alias{aggregations}
\alias{aggregations.GDCQuery}
\alias{aggregations.GDCResponse}
\title{aggregations}
\usage{
aggregations(x)

\method{aggregations}{GDCQuery}(x)

\method{aggregations}{GDCResponse}(x)
}
\arguments{
\item{x}{a \code{\link{GDCQuery}} object}
}
\description{
aggregations
}
\section{Methods (by class)}{
\itemize{
\item \code{GDCQuery}: 

\item \code{GDCResponse}: 
}}
\examples{
# Number of each file type
res = files() \%>\% facet(c('type','data_type')) \%>\% aggregations()
res$type

}

