% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotations}
\alias{annotations}
\title{Query GDC for annotations}
\usage{
annotations(..., fields = annotation_fields())
}
\arguments{
\item{fields}{character() vector of requested fields. See
\code{annotation_fields()} for defined fields.}

\item{\dots}{(optional) additional parameters influence annotation
selection. See \code{\link{parameters}}.}
}
\description{
Query GDC for annotations
}
\examples{
annotations <- annotations()
annotations
annotations[1]
unlist(annotations[[1]])

}

