% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_results.R
\name{as.data.frame.GDCResults}
\alias{as.data.frame.GDCResults}
\title{Convert GDC results to data.frame}
\usage{
\method{as.data.frame}{GDCResults}(res)
}
\arguments{
\item{res}{a \code{GDCResults} object to be converted to a 
data.frame.}
}
\value{
a data.frame, potentially with list columns still
    present.
}
\description{
GDC results are typically returned as an R list
structure. This method converts that 
R list structure to a data.frame. Some columns
in the resulting data.frame may remain lists, but
there will be one list element for each row though
that list element may contain multiple values.
}
\examples{
expands = c("diagnoses","diagnoses.treatments","annotations",
            "demographic","exposures")
head(cases() \%>\% expand(expands) \%>\% as.data.frame())

}
