% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{available_fields}
\alias{available_fields}
\alias{available_fields.GDCQuery}
\alias{available_fields.character}
\title{S3 Generic to return all GDC fields}
\usage{
available_fields(x)

\method{available_fields}{GDCQuery}(x)

\method{available_fields}{character}(x)
}
\arguments{
\item{x}{A character(1) string ('cases','files','projects',
'annotations') or an subclass of \code{\link{GDCQuery}}.}
}
\value{
a character vector of the default fields
}
\description{
S3 Generic to return all GDC fields
}
\section{Methods (by class)}{
\itemize{
\item \code{GDCQuery}: GDCQuery method

\item \code{character}: character method
}}
\examples{
available_fields('projects')
projQuery = query('projects')
available_fields(projQuery)

}

