% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{available_values}
\alias{available_values}
\title{Find common values for a GDC field}
\usage{
available_values(entity, field, legacy = FALSE)
}
\arguments{
\item{entity}{character(1), a GDC entity ("cases", "files", "annotations", "projects")}

\item{field}{character(1), a field that is present in the entity record}

\item{legacy}{logical(1), DEPRECATED; use the legacy endpoint or not.}
}
\value{
character vector of the top 100 (or fewer) most frequent
    values for a the given field
}
\description{
Find common values for a GDC field
}
\examples{
available_values('files','cases.project.project_id')[1:5]

}
