% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{case_fields}
\alias{case_fields}
\alias{cases}
\title{Query GDC for cases}
\usage{
case_fields(primary = TRUE)

cases(..., token = NULL, fields = case_fields())
}
\arguments{
\item{primary}{logical(1) when TRUE (default) return commonly
populated field names. Otherwise, return all field names
defined in the API.}

\item{token}{(optional) character(1) security token allowing access
to restricted data. See
\url{https://gdc-docs.nci.nih.gov/API/Users_Guide/Authentication_and_Authorization/}.}

\item{fields}{character() vector of requested fields. See
\code{case_fields()} for defined fields.}

\item{\dots}{(optional) additional parameters influence case
selection. See \code{\link{parameters}}.}
}
\description{
Query GDC for cases
}
\examples{
cases <- cases()
cases
cases[1]
cases[[1]]$sample_ids

lst <- cases(fields="submitter_id", sort="submitter_id:asc")
lst
sapply(lst, "[[", "submitter_id")

}

