% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{endpoints}
\alias{endpoints}
\alias{parameters}
\title{Endpoints and Parameters}
\usage{
endpoints()

parameters()
}
\value{
\code{endpoints()} returns a character vector of possible
    endpoints.

\code{parameters()} returns a list of possible parameters
    and their default values.
}
\description{
\code{endpoints()} returns available endpoints.
}
\examples{
endpoints()
parameters()
}
\keyword{internal}
