% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{count}
\alias{count}
\alias{count.GDCQuery}
\title{provide count of records in a GDCQuery}
\usage{
count(x, ...)

\method{count}{GDCQuery}(x, ...)
}
\arguments{
\item{x}{a \code{\link{GDCQuery}} object}

\item{...}{passed to httr (good for passing config info, etc.)}
}
\description{
provide count of records in a GDCQuery
}
\section{Methods (by class)}{
\itemize{
\item \code{GDCQuery}: 
}}

