% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{default_fields}
\alias{default_fields}
\alias{default_fields.character}
\alias{default_fields.GDCQuery}
\title{S3 Generic to return default GDC fields}
\usage{
default_fields(x)

\method{default_fields}{character}(x)

\method{default_fields}{GDCQuery}(x)
}
\arguments{
\item{x}{A character string ('cases','files','projects',
'annotations') or an subclass of \code{\link{GDCQuery}}.}
}
\value{
a character vector of the default fields
}
\description{
S3 Generic to return default GDC fields
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: character method

\item \code{GDCQuery}: GDCQuery method
}}

\examples{
default_fields('projects')
projQuery = query('projects')
default_fields(projQuery)

}
