% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{field_picker}
\alias{field_picker}
\title{Interactive shiny gadget to pick metadata fields}
\usage{
field_picker(entity)
}
\arguments{
\item{entity}{One of the metadata entities or a GDCQuery}
}
\value{
a character vector of chosen fields. The entity
id field (eg., 'case_id' for cases) will be automatically
included.
}
\description{
Since each of the metadata endpoints can have dozens or
even hundreds of fields, having an approach to quickly
find and choose these fields interactively is really
useful. When running this function, a browser window
will pop up or, if in RStudio, an interactive widget
will show up in the viewer pane. Type and choose your
fields of interest. When complete, the fields will
be returned to the R session for capture in a variable.
}
\examples{
\dontrun{
fchoices = field_picker('projects')
fchoices
q = projects() \%>\% select(fchoices)
results(q)[[1]]

# or inline
q = projects() \%>\% select(field_picker('projects'))
head(q$fields)
}
}

