% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_fields}
\alias{file_fields}
\alias{files}
\title{Query GDC for files}
\usage{
file_fields(primary = TRUE)

files(..., fields = file_fields())
}
\arguments{
\item{primary}{logical(1) when TRUE (default) return commonly
populated field names. Otherwise, return all field names
defined in the API.}

\item{fields}{character() vector of requested fields. See
\code{file_fields()} for defined fields.}

\item{\dots}{(optional) additional parameters influence file
selection. See \code{\link{parameters}}.}
}
\description{
Query GDC for files
}
\examples{
files <- files()
files
files[1]
unlist(files[[1]])

}

