% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_expr}
\alias{filter_expr}
\title{Create GDC filters}
\usage{
filter_expr(expr, available_fields, endpoint)
}
\arguments{
\item{endpoint}{one of the recognized endpoints; used to pull accepted fields for
use in query construction}

\item{filter_expression}{an R-like expression}

\item{...}{passed to \code{jsonlite::toJSON}}
}
\value{
either a JSON string or the R list prior to conversion to JSON
}
\description{
Create filters using standard R formatting
}
\details{
This function makes arbitrarily complex filter functions for
and creates a list that is ready for easy conversion with
toJSON for use in queries of the GDC.
}
\examples{
#filter_expr(diagnoses.age_at_diagnosis <= 10*365,endpoint='cases')
#filter_expr(project.primary_site \%in\% c('blood','brain'),endpoint='cases')
}

