% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filters}
\alias{filters}
\title{Create GDC filters}
\usage{
filters(filter_expression, endpoint, asJSON = TRUE, ...)
}
\arguments{
\item{filter_expression}{an R expression}

\item{asJSON}{if TRUE, return the JSON.  If FALSE, return an R list}

\item{...}{passed to \code{jsonlite::toJSON}}
}
\value{
either a JSON string or the R list prior to conversion to JSON
}
\description{
Create filters using standard R formatting
}
\details{
This function makes arbitrarily complex filter functions for
and creates translated JSON for use in queries of the GDC.
}
\examples{
filters(cases.clinical.age_at_diagnosis >= 14600 & cases.clinical.age_at_diagnosis <= 25500,endpoint='cases')
filters(primary_site \%in\% c('blood','brain'),endpoint='cases')
}

