% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDC-defunct.R
\name{GDC-defunct}
\alias{GDC-defunct}
\alias{legacy}
\alias{legacy.GDCQuery}
\title{Defunct functionality in the `GenomicDataCommons` package}
\usage{
legacy(x, legacy)

\method{legacy}{GDCQuery}(x, legacy)
}
\arguments{
\item{x}{the objects on which to set fields}

\item{legacy}{logical(1) DEFUNCT; no longer in use.}
}
\value{
An error. This feature is no longer used.
}
\description{
The `legacy` endpoint was removed after the GDC API release version
`v3.28.0`. This means that legacy data is no longer available via the
GDC API. The `legacy` features in the `GenomicDataCommons` package are
defunct.
}
\details{
This includes the S3 generic and method for setting the legacy parameter
(`legacy` and `legacy.GDCQuery`)
}
\section{Functions}{
\itemize{
\item \code{legacy(GDCQuery)}: set legacy field on a GDCQuery object

}}
\examples{
qcases <- query("cases")
}
\seealso{
\url{https://docs.gdc.cancer.gov/API/Release_Notes/API_Release_Notes/#v3280}
}
