% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDC-deprecated.R
\name{GDC-deprecated}
\alias{GDC-deprecated}
\alias{legacy}
\alias{legacy.GDCQuery}
\title{Deprecated functionality in the `GenomicDataCommons` package}
\usage{
legacy(x, legacy)

\method{legacy}{GDCQuery}(x, legacy)
}
\arguments{
\item{x}{the objects on which to set fields}

\item{legacy}{logical(1) DEPRECATED; Whether or not to use the GDC legacy
archives}
}
\value{
A \code{\link{GDCQuery}} object with the \code{legacy}
member altered.
}
\description{
The `legacy` endpoint was deprecated as part of the GDC API release version
`v3.28.0`. This means that legacy data is no longer available via the
GDC API.
}
\details{
This includes the S3 generic and method for setting the legacy parameter
(`legacy` and `legacy.GDCQuery`)
}
\section{Functions}{
\itemize{
\item \code{legacy(GDCQuery)}: set legacy field on a GDCQuery object

}}
\examples{
qcases <- query("cases")

legacy(qcases, legacy = FALSE)

}
\seealso{
\url{https://docs.gdc.cancer.gov/API/Release_Notes/API_Release_Notes/#v3280}
}
