% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{gdc_cache}
\alias{gdc_cache}
\alias{gdc_set_cache}
\title{Work with gdc cache directory}
\usage{
gdc_cache(directory = NULL)

gdc_set_cache(directory)
}
\arguments{
\item{directory}{character(1) directory path, will be
created recursively if not present.}
}
\value{
character(1) directory path that serves as
the base directory for GenomicDataCommons downloads.
}
\description{
The GenomicDataCommons package will cache downloaded
files to minimize network and allow for
offline work. These functions
}
\section{Functions}{
\itemize{
\item \code{gdc_set_cache}: (Re)set the GenomicDataCommons cache directory
}}

\examples{
gdc_cache()
\dontrun{
gdc_set_cache(getwd())
}

}
