% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{gdcAvailableFields}
\alias{gdcAvailableFields}
\alias{gdcAvailableFields.GDCQuery}
\alias{gdcAvailableFields.character}
\title{S3 Generic to return all GDC fields}
\usage{
gdcAvailableFields(x)

\method{gdcAvailableFields}{GDCQuery}(x)

\method{gdcAvailableFields}{character}(x)
}
\arguments{
\item{x}{A character string ('cases','files','projects',
'annotations') or an subclass of \code{\link{GDCQuery}}.}
}
\value{
a character vector of the default fields
}
\description{
S3 Generic to return all GDC fields
}
\section{Methods (by class)}{
\itemize{
\item \code{GDCQuery}: GDCQuery method

\item \code{character}: character method
}}
\examples{
gdcAvailableFields('projects')
projQuery = gdcQuery('projects')
gdcAvailableFields(projQuery)

}

