% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcdata.R
\name{gdcdata}
\alias{gdcdata}
\title{Download GDC files}
\usage{
gdcdata(uuids, destination_dir = tempfile(), overwrite = FALSE,
  progress = interactive())
}
\arguments{
\item{uuids}{character() of GDC file UUIDs.}

\item{destination_dir}{character(1) file path to a directory for
downloading files.}

\item{overwrite}{logical(1) default FALSE indicating whether
existing files with identical name should be over-written.}

\item{progress}{logical(1) default TRUE in interactive sessions,
FALSE otherwise indicating whether a progress par should be
produced for each file download}
}
\description{
Download one or more files from GDC. Files are downloaded using the
UUID and renamed to the file name on the remote system. By default,
neither the uuid nor the file name on the remote system can exist.
}
\details{
This function is appropriate for one or several files; for large
downloads use \code{\link{manifest}} to create a manifest for and
the GDC Data Transfer Tool.
}
\examples{
uuids <- c("e3228020-1c54-4521-9182-1ea14c5dc0f7",
           "18e1e38e-0f0a-4a0e-918f-08e6201ea140")
(files <- gdcdata(uuids, overwrite=TRUE))
setNames(file.size(files), names(files))
}
\seealso{
\code{\link{manifest}} for downloading large data.
}

