% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{gdcDefaultFields}
\alias{gdcDefaultFields}
\alias{gdcDefaultFields.GDCQuery}
\alias{gdcDefaultFields.character}
\title{S3 Generic to return default GDC fields}
\usage{
gdcDefaultFields(x)

\method{gdcDefaultFields}{character}(x)

\method{gdcDefaultFields}{GDCQuery}(x)
}
\arguments{
\item{x}{A character string ('cases','files','projects',
'annotations') or an subclass of \code{\link{gdcQuery-class}}.}
}
\value{
a character vector of the default fields
}
\description{
S3 Generic to return default GDC fields
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: character method

\item \code{GDCQuery}: GDCQuery method
}}
\examples{
gdcDefaultFields('projects')
projQuery = gdcQuery('projects')
gdcDefaultFields(projQuery)

}

