% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{gdcEntityName}
\alias{gdcEntityName}
\title{Get the entity name from a GDCQuery object}
\usage{
gdcEntityName(x)
}
\arguments{
\item{x}{a \code{\link{GDCQuery}} object}
}
\description{
Get the entity name from a GDCQuery object
}

