% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{gdcQuery}
\alias{gdcAannotations}
\alias{gdcCases}
\alias{gdcEntityName}
\alias{gdcFiles}
\alias{gdcProjects}
\alias{gdcQuery}
\title{Construct a GDCQuery}
\usage{
gdcQuery(entity, token = NULL, filters = NULL, facets = NULL,
  fields = subset(mapping(entity), defaults)$field)

gdcCases(...)

gdcFiles(...)

gdcProjects(...)

gdcAannotations(...)

gdcEntityName(gdcQuery)
}
\arguments{
\item{token}{a character string reprenting the token}

\item{filters}{a filter list}

\item{facets}{a facets list}

\item{fields}{a list of fields to return}
}
\value{
An S3 object, the GDCQuery object. This is a list
with filters, facets, and fields as members.
}
\description{
The basis for all functionality in this package
starts with constructing a query in R. The GDCQuery
object contains the filters, facets, and other
parameters that define the returned results. A token
is required for accessing certain datasets.
}
\section{Functions}{
\itemize{
\item \code{gdcCases}: convenience contructor for a GDCQuery for cases

\item \code{gdcFiles}: convenience contructor for a GDCQuery for cases

\item \code{gdcProjects}: convenience contructor for a GDCQuery for cases

\item \code{gdcAannotations}: convenience contructor for a GDCQuery for annotations

\item \code{gdcEntityName}: get the entity name from a GDCQuery object
}}

