% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicDataCommons-package.R
\docType{package}
\name{GenomicDataCommons-package}
\alias{GenomicDataCommons}
\alias{GenomicDataCommons-package}
\title{GenomicDataCommons: A package for interfacing with the NCI GDC}
\description{
Programmatically access the NIH / NCI Genomic Data Commons RESTful service.
}
\section{finding data}{


\itemize{
\item{\code{\link{query}}}
\item{\code{\link{cases}}}
\item{\code{\link{projects}}}
\item{\code{\link{files}}}
\item{\code{\link{annotations}}}
\item{\code{\link{mapping}}}
}
}

\section{downloading data}{

data
}

\seealso{
Useful links:
\itemize{
  \item \url{https://bioconductor.org/packages/GenomicDataCommons}
  \item \url{http://github.com/Bioconductor/GenomicDataCommons}
  \item \url{http://bioconductor.github.io/GenomicDataCommons/}
  \item Report bugs at \url{https://github.com/Bioconductor/GenomicDataCommons/issues/new}
}

}
\author{
\strong{Maintainer}: Sean Davis \email{seandavi@gmail.com}

Authors:
\itemize{
  \item Martin Morgan \email{martin.morgan@roswellpark.org}
}

Other contributors:
\itemize{
  \item Marcel Ramos \email{marcel.ramos@sph.cuny.edu} [contributor]
}

}
