% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{id_field}
\alias{id_field}
\alias{id_field.GDCQuery}
\alias{id_field.GDCResults}
\title{get the name of the id field}
\usage{
id_field(x)

\method{id_field}{GDCQuery}(x)

\method{id_field}{GDCResults}(x)
}
\arguments{
\item{x}{An object representing the query or results 
of an entity from the GDC ("cases", "files", "annotations", "projects")}
}
\value{
character(1) such as "case_id", "file_id", etc.
}
\description{
In many places in the GenomicDataCommons package,
the entity ids are stored in a column or a vector
with a specific name that corresponds to the field name 
at the GDC. The format is the entity name (singular) "_id".
This generic simply returns that name from a given object.
}
\section{Methods (by class)}{
\itemize{
\item \code{GDCQuery}: GDCQuery method

\item \code{GDCResults}: GDCResults method
}}
\examples{
id_field(cases())

}

