% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{legacy}
\alias{legacy}
\alias{legacy.GDCQuery}
\title{Set the \code{legacy} parameter}
\usage{
legacy(x, legacy)

\method{legacy}{GDCQuery}(x, legacy)
}
\arguments{
\item{x}{the objects on which to set fields}

\item{legacy}{logical(1) Whether or not to use the GDC legacy archives}
}
\value{
A \code{\link{GDCQuery}} object with the \code{legacy}
member altered.
}
\description{
S3 generic to set GDCQuery legacy parameter
}
\section{Methods (by class)}{
\itemize{
\item \code{legacy(GDCQuery)}: set legacy field on a GDCQuery object

}}
\examples{
qcases <- query("cases")

legacy(qcases, legacy = FALSE)

}
