% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{make_filter2}
\alias{make_filter2}
\title{makefilter2}
\usage{
make_filter2(expr, available_fields)
}
\arguments{
\item{expr}{a filter expression}

\item{available_fields}{a character vector of the
additional names that will be injected into the
filter evaluation environment}
}
\description{
makefilter2
}

