% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest}
\alias{manifest}
\alias{transfer}
\alias{transfer_help}
\title{GDC manifest file creation and file download}
\usage{
manifest(uuids, destination_dir = tempfile())

transfer(manifest, destination_dir = tempfile(), args = character(),
  gdc_client = "gdc-client")

transfer_help(gdc_client = "gdc-client")
}
\arguments{
\item{uuids}{character() of GDC file UUIDs.}

\item{destination_dir}{character(1) file path to a directory for
downloading the manifest or files. The manifest file name, of
the form \sQuote{gdc_manifest_20160610_12345.txt}, must not
exist.}

\item{manifest}{character(1) file path to manifest created by
\code{manifest()}}

\item{args}{character() vector specifying command-line arguments to
be passed to \code{gdc-client}. See \code{transfer_help} for
possible values. The arguments \code{--manifest} and
\code{--dir} are determined by \code{manifest} and
\code{destination_dir}, respectively, and should NOT be
provided.}

\item{gdc_client}{character(1) name or path to \code{gdc-client}
executable. On Windows, use \code{/} or \code{\\\\} as the file
separator.}
}
\value{
character(1) directory path to which the files were
    downloaded.
}
\description{
\code{manifest()} creates a manifest of files to be downloaded
using the GDC Data Transfer Tool.

\code{transfer()} retrieves files from the manifest using the GDC
Data Transfer Tool.

\code{transfer_help()} queries the the command line GDC Data
Transfer Tool, \code{gdc-client}, for available options to be used
in the \code{transfer()} command.
}
\examples{
uuids <- c("e3228020-1c54-4521-9182-1ea14c5dc0f7",
           "18e1e38e-0f0a-4a0e-918f-08e6201ea140")
(manifest <- manifest(uuids))
\donttest{
destination <- transfer(manifest)
dir(destination)
}
}

