% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping}
\alias{mapping}
\title{Query GDC for available endpoint fields}
\usage{
mapping(endpoint, token = NULL)
}
\arguments{
\item{endpoint}{character(1) corresponding to endpoints for which
users may specify additional or alternative fields. Endpoints
include \dQuote{projects}, \dQuote{cases}, \dQuote{files}, and
\dQuote{annotations}.}

\item{token}{(optional) character(1) security token allowing access
to restricted data. See
\url{https://gdc-docs.nci.nih.gov/API/Users_Guide/Authentication_and_Authorization/}.}
}
\description{
Query GDC for available endpoint fields
}
\examples{
map <- mapping("projects")
map["default"]
}

