% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects}
\alias{projects}
\title{Query GDC for projects}
\usage{
projects(..., fields = project_fields())
}
\arguments{
\item{\dots}{(optional) additional parameters influencing project
selection. See \code{\link{parameters}}.}
}
\description{
Query GDC for projects
}
\examples{
projects()                        # first 10 projects, as a data.frame
df <- projects(from=10, size=20)  # projects 10, 11, ... 29
dim(df)

}

