% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_results.R
\name{rbindlist2}
\alias{rbindlist2}
\title{rbindlist that deals with \code{NULL} rows}
\usage{
rbindlist2(x, ...)
}
\arguments{
\item{x}{a list that is appropriate for input
to \code{\link[data.table]{rbindlist}}, but
may include \code{NULL}s, which will be filtered.}

\item{...}{passed directly to \code{\link[data.table]{rbindlist}}.}
}
\value{
a \code{data.table,data.frame} object.
}
\description{
This is a simple function that removes \code{null}
values from the input list
before applying \code{\link[data.table]{rbindlist}}.
}
\examples{
input = list(list(a=1,b=2,d='only this row'),
             NULL,
             list(a=3,b=4))
rbindlist2(input)

}
