% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{results}
\alias{results}
\alias{results.GDCQuery}
\alias{results.GDCResponse}
\title{results}
\usage{
results(x)

results(x)

\method{results}{GDCQuery}(x)

\method{results}{GDCResponse}(x)
}
\arguments{
\item{x}{a \code{\link{GDCQuery}} object}

\item{x}{a \code{\link{GDCQuery}} object}
}
\description{
results

results_all
}
\section{Methods (by class)}{
\itemize{
\item \code{GDCQuery}: 

\item \code{GDCResponse}: 
}}

