% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{results_all}
\alias{results_all}
\alias{results_all.GDCQuery}
\alias{results_all.GDCResponse}
\title{results_all}
\usage{
results_all(x)

\method{results_all}{GDCQuery}(x)

\method{results_all}{GDCResponse}(x)
}
\arguments{
\item{x}{a \code{\link{GDCQuery}} object}
}
\value{
A (typically nested) \code{list} of GDC records
}
\description{
results_all
}
\section{Methods (by class)}{
\itemize{
\item \code{GDCQuery}: 

\item \code{GDCResponse}: 
}}

\examples{
# details of all available projects
projResults = projects() \%>\% results_all()
length(projResults)
count(projects())


}
