% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{Query the GDC for current status}
\usage{
status(version = NULL)
}
\arguments{
\item{version}{(optional) character(1) version of GDC}
}
\value{
List describing current status.
}
\description{
Query the GDC for current status
}

