% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_transfer.R
\name{transfer_help}
\alias{transfer_help}
\title{\code{transfer_help()} queries the the command line GDC Data
Transfer Tool, \code{gdc-client}, for available options to be used
in the \code{\link{transfer}} command.}
\usage{
transfer_help(gdc_client = "gdc-client")
}
\description{
\code{transfer_help()} queries the the command line GDC Data
Transfer Tool, \code{gdc-client}, for available options to be used
in the \code{\link{transfer}} command.
}

