% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotation_fields}
\alias{annotation_fields}
\alias{annotations}
\title{Query GDC for annotations}
\usage{
annotation_fields(primary = TRUE)

annotations(..., fields = annotation_fields())
}
\arguments{
\item{primary}{logical(1) when TRUE (default) return commonly
populated field names. Otherwise, return all field names
defined in the API.}

\item{fields}{character() vector of requested fields. See
\code{annotation_fields()} for defined fields.}

\item{\dots}{(optional) additional parameters influence annotation
selection. See \code{\link{parameters}}.}
}
\description{
Query GDC for annotations
}
\examples{
annotations <- annotations()
annotations
annotations[1]
unlist(annotations[[1]])

}

