% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_results.R
\name{as.data.frame.GDCResults}
\alias{as.data.frame.GDCResults}
\title{Convert GDC results to data.frame}
\usage{
\method{as.data.frame}{GDCResults}(x, row.names, optional, ...)
}
\arguments{
\item{x}{a \code{GDCResults} object to be converted to a 
data.frame.}

\item{row.names}{not used here; row.names are calculated
from the data to maintain data integrity}

\item{optional}{not used here; just for matching method call for \code{\link{as.data.frame}}}

\item{...}{not used here; just for matching method call for \code{\link{as.data.frame}}}
}
\value{
a data.frame, potentially with list columns still
    present.
}
\description{
GDC results are typically returned as an R list
structure. This method converts that 
R list structure to a data.frame. Some columns
in the resulting data.frame may remain lists, but
there will be one list element for each row though
that list element may contain multiple values.
}
\note{
The data.frame that is returned is the cartesian product 
of sub-data.frames that come from the GDC. This may lead to more
than one row per entity id. It is up to the user to determine 
how to make rows unique per entity, if that is desired.
}
\examples{
expands = c("diagnoses","diagnoses.treatments","annotations",
            "demographic","exposures")
head(cases() \%>\% expand(expands) \%>\% results() \%>\% as.data.frame())

}

