% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R, R/files.R
\name{case_fields}
\alias{case_fields}
\alias{cases}
\alias{file_fields}
\title{Query GDC for cases}
\usage{
case_fields(primary = TRUE)

cases(..., fields = case_fields())

file_fields(primary = TRUE)
}
\arguments{
\item{primary}{logical(1) when TRUE (default) return commonly
populated field names. Otherwise, return all field names
defined in the API.}

\item{fields}{character() vector of requested fields. See
\code{case_fields()} for defined fields.}

\item{\dots}{(optional) additional parameters influence case
selection. See \code{\link{parameters}}.}

\item{primary}{logical(1) when TRUE (default) return commonly
populated field names. Otherwise, return all field names
defined in the API.}
}
\description{
Query GDC for cases
}
\examples{
cases <- cases()
cases
cases[1]
cases[[1]]$sample_ids

}

