% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{count}
\alias{count}
\alias{count.GDCQuery}
\alias{count.GDCResponse}
\title{provide count of records in a \code{\link{GDCQuery}}}
\usage{
count(x, ...)

\method{count}{GDCQuery}(x, ...)

\method{count}{GDCResponse}(x, ...)
}
\arguments{
\item{x}{a \code{\link{GDCQuery}} object}

\item{...}{passed to httr (good for passing config info, etc.)}
}
\description{
provide count of records in a \code{\link{GDCQuery}}
}
\section{Methods (by class)}{
\itemize{
\item \code{GDCQuery}: 

\item \code{GDCResponse}: 
}}

