% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{entity_name}
\alias{entity_name}
\title{Get the entity name from a GDCQuery object}
\usage{
entity_name(x)
}
\arguments{
\item{x}{a \code{\link{GDCQuery}} object}
}
\description{
Get the entity name from a GDCQuery object
}

