% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets.R
\name{gdcSetFacet}
\alias{gdcGetFacet}
\alias{gdcGetFacet.GDCQuery}
\alias{gdcSetFacet}
\title{Set facets for a \code{\link{GDCQuery}}}
\usage{
gdcSetFacet(x, facets)

gdcGetFacet(x)

\method{gdcGetFacet}{GDCQuery}(x)
}
\arguments{
\item{x}{a \code{\link{GDCQuery}} object}

\item{facets}{a character vector of fields that
will be used for forming aggregations (facets).
Default is to set facets for all default fields.
See \code{\link{default_fields}} for details}
}
\description{
Set facets for a \code{\link{GDCQuery}}

Get facets for a \code{\link{GDCQuery}}
}

