% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files}
\alias{files}
\title{Query GDC for files}
\usage{
files(..., fields = file_fields())
}
\arguments{
\item{fields}{character() vector of requested fields. See
\code{file_fields()} for defined fields.}

\item{\dots}{(optional) additional parameters influence file
selection. See \code{\link{parameters}}.}
}
\description{
Query GDC for files
}
\examples{
files <- files()
files
files[1]
unlist(files[[1]])

}

