% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_expr}
\alias{filter_expr}
\title{Create NCI GDC filters for limiting GDC query results}
\usage{
filter_expr(expr, available_fields = NULL)
}
\arguments{
\item{expr}{an R expression}

\item{available_fields}{A character vector of field names that, if specified,
can be used unquoted in creating filter expressions}
}
\value{
a (potentially nested) R list prior that, after converting to JSON
can be used as the filter parameter in an NCI GDC search or other query.
}
\description{
Create filters using standard R formatting
}
\details{
Searching the NCI GDC allows for complex filtering based
on logical operations and simple comparisons.  This function
facilitates writing such filter expressions in R-like syntax
with R code evaluation.
}
\examples{
filter_expr("diagnoses.age_at_diagnosis" <= 10*365)
filter_expr("project.primary_site" \%in\% c('blood','brain'))
fe = filter_expr("project.primary_site" \%in\% c('blood','brain')
                 & "diagnosis.age_at_diagnosis" <= 10*365)
jsonlite::toJSON(fe,auto_unbox=TRUE,pretty=TRUE)

}

