% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{gdc_client}
\alias{gdc_client}
\title{return gdc-client executable path}
\usage{
gdc_client()
}
\value{
character(1) the path to the gdc-client executable.
}
\description{
This function is a convenience function to 
find and return the path to the GDC Data Transfer
Tool executable assumed to be named 'gdc-client'.
The following locations are checked:
}
\details{
\itemize{
\item{Sys.which() to see if gdc-client is on the path}
\item{The current working directory}
\item{The file name specified in the environment variable \code{GDC_CLIENT}}
}
}
