% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{gdcSetFields}
\alias{gdcSetFields}
\alias{gdcSetFields.GDCQuery}
\title{S3 generic to set GDCQuery fields}
\usage{
gdcSetFields(x, fields)

\method{gdcSetFields}{GDCQuery}(x, fields)
}
\description{
S3 generic to set GDCQuery fields
}
\section{Methods (by class)}{
\itemize{
\item \code{GDCQuery}: set fields on a GDCQuery object
}}

