% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{grep_fields}
\alias{grep_fields}
\title{Find matching field names}
\usage{
grep_fields(entity, pattern, ..., value = TRUE)
}
\arguments{
\item{entity}{one of "files", "cases", "annotations", "projects"
against which to gather available fields for matching}

\item{pattern}{A regular expression that will be used
in a call to \code{\link{grep}}}

\item{...}{passed on to grep}

\item{value}{logical(1) whether to return values as opposed
to indices (passed along to grep)}
}
\value{
character() vector of field names matching
    \code{pattern}
}
\description{
This utility function allows quick text-based search of available
fields for using \code{\link{grep}}
}
\examples{
grep_fields('files','analysis')

}

