% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{make_filter}
\alias{make_filter}
\title{Create NCI GDC filters for limiting GDC query results}
\usage{
make_filter(expr, available_fields)
}
\arguments{
\item{expr}{a filter expression}

\item{available_fields}{a character vector of the
additional names that will be injected into the
filter evaluation environment}
}
\value{
a \code{list} that represents an R version
of the JSON that will ultimately be used in an
NCI GDC search or other query.
}
\description{
Searching the NCI GDC allows for complex filtering based
on logical operations and simple comparisons.  This function
facilitates writing such filter expressions in R-like syntax
with R code evaluation.
}
\details{
If used with available_fields, "bare" fields that are
named in the available_fields character vector can be used
in the filter expression without quotes.
}
