% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest}
\alias{manifest}
\alias{manifest.GDCcasesResponse}
\alias{manifest.GDCfilesResponse}
\alias{manifest.gdc_cases}
\alias{manifest.gdc_files}
\alias{transfer}
\alias{transfer_help}
\title{GDC manifest file creation and file download}
\usage{
manifest(x, from = 1, size = count(x), ...)

\method{manifest}{gdc_files}(x, from = 1, size = count(x), ...)

\method{manifest}{gdc_cases}(x, from = 1, size = count(x), ...)

\method{manifest}{GDCfilesResponse}(x, from = 1, size = count(x), ...)

\method{manifest}{GDCcasesResponse}(x, from = 1, size = count(x), ...)

transfer(manifest, destination_dir = tempfile(), args = character(),
  token_file = NULL, gdc_client = "gdc-client")

transfer_help(gdc_client = "gdc-client")
}
\arguments{
\item{x}{An \code{\link{GDCQuery}} object of type 'gdc_files'.}

\item{from}{Record number from which to start when returning the manifest.}

\item{size}{The total number of records to return.  Default will return the usually desirable full set of records.}

\item{...}{passed to \code{\link[httr]{PUT}}.}

\item{manifest}{character(1) file path to manifest created by
\code{manifest()}}

\item{destination_dir}{The path into which to place the transfered files.}

\item{args}{character() vector specifying command-line arguments to
be passed to \code{gdc-client}. See \code{transfer_help} for
possible values. The arguments \code{--manifest}, \code{--dir},
and \code{--token-file} are determined by \code{manifest},
\code{destination_dir}, and \code{token}, respectively, and
should NOT be provided as elements of \code{args}.}

\item{token_file}{character(1) path to a file containing security
token allowing access to restricted data. See
\url{https://gdc-docs.nci.nih.gov/API/Users_Guide/Authentication_and_Authorization/}.}

\item{gdc_client}{character(1) name or path to \code{gdc-client}
executable. On Windows, use \code{/} or \code{\\\\} as the file
separator.}
}
\value{
A \code{\link[tibble]{tibble}} with five columns:
\itemize{
\item{id}
\item{filename}
\item{md5}
\item{size}
\item{state}
}

character(1) directory path to which the files were
    downloaded.
}
\description{
\code{manifest()} creates a manifest of files to be downloaded
using the GDC Data Transfer Tool.

\code{transfer()} retrieves files from the manifest using the GDC
Data Transfer Tool.

\code{transfer_help()} queries the the command line GDC Data
Transfer Tool, \code{gdc-client}, for available options to be used
in the \code{transfer()} command.
}
\section{Methods (by class)}{
\itemize{
\item \code{gdc_files}: 

\item \code{gdc_cases}: 

\item \code{GDCfilesResponse}: 

\item \code{GDCcasesResponse}: 
}}
\examples{
gFiles = files()
shortManifest = gFiles \%>\% manifest(size=10)
head(shortManifest,n=3)

gCases = cases()
manifestFromCases = manifest(gCases,size=10)
manifestFromCases


\donttest{
destination <- transfer(manifest)
dir(destination)
}
}

