% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{project_fields}
\alias{project_fields}
\alias{projects}
\title{Query GDC for projects}
\usage{
project_fields(primary = TRUE)

projects(..., fields = project_fields())
}
\arguments{
\item{primary}{logical(1) when TRUE (default) return commonly
populated field names. Otherwise, return all field names
defined in the API.}

\item{fields}{character() vector of requested fields. See
\code{project_fields()} for defined fields.}

\item{\dots}{(optional) additional parameters influencing project
selection. See \code{\link{parameters}}.}
}
\description{
Query GDC for projects
}
\examples{
projects()                        # first 10 projects, as a data.frame
df <- projects(from=10, size=20)  # projects 10, 11, ... 29
dim(df)

}

