% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDNAcopy.R
\name{readDNAcopy}
\alias{readDNAcopy}
\title{Read DNAcopy results into GRanges object}
\usage{
readDNAcopy(fname, ...)
}
\arguments{
\item{fname}{The path to a DNAcopy-like file.}
}
\value{
a \code{\link[GenomicRanges]{GRanges}} object
}
\description{
Read DNAcopy results into GRanges object
}
\examples{
fname = system.file(package='GenomicDataCommons',
                    'extdata/dnacopy.tsv.gz')
dnac = readDNAcopy(fname)
class(dnac)
seqnames(dnac)
length(dnac)
}
