% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{response}
\alias{response}
\alias{response.GDCQuery}
\alias{response_all}
\title{get the response from server}
\usage{
response(x, ...)

\method{response}{GDCQuery}(x, from = 1, size = 10, ...)

response_all(x, ...)
}
\arguments{
\item{x}{a \code{\link{GDCQuery}} object}

\item{...}{passed to httr (good for passing config info, etc.)}

\item{from}{integer index from which to start returning data}

\item{size}{number of records to return}
}
\description{
get the response from server
}

