% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{results}
\alias{results}
\alias{results.GDCQuery}
\alias{results.GDCResponse}
\title{results}
\usage{
results(x, ...)

\method{results}{GDCQuery}(x, ...)

\method{results}{GDCResponse}(x, ...)
}
\arguments{
\item{x}{a \code{\link{GDCQuery}} object}

\item{...}{passed on to \code{\link{response}}}
}
\value{
A (typically nested) \code{list} of GDC records
}
\description{
results
}
\section{Methods (by class)}{
\itemize{
\item \code{results(GDCQuery)}: 

\item \code{results(GDCResponse)}: 

}}
\examples{
qcases = cases() |> results()
length(qcases)

}
