% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_transfer.R
\name{transfer}
\alias{transfer}
\alias{transfer_help}
\title{Bulk data download}
\usage{
transfer(uuids, args = character(), token = NULL, overwrite = FALSE)

transfer_help()
}
\arguments{
\item{uuids}{character() vector of GDC file UUIDs}

\item{args}{character() vector specifying command-line arguments to
be passed to \code{gdc-client}. See \code{\link{transfer_help}} for
possible values. The arguments \code{--manifest}, \code{--dir},
and \code{--token-file} are determined by \code{manifest},
\code{destination_dir}, and \code{token}, respectively, and
should NOT be provided as elements of \code{args}.}

\item{token}{character(1) containing security
token allowing access to restricted data. See
\url{https://gdc-docs.nci.nih.gov/API/Users_Guide/Authentication_and_Authorization/}.
Note that the GDC transfer tool requires a file for data
transfer. Therefore, this token will be written to a temporary
file (with appropriate permissions set).}

\item{overwrite}{logical(1) default FALSE indicating whether
existing files with identical name should be over-written.}
}
\value{
character(1) directory path to which the files were
    downloaded.
}
\description{
The GDC maintains a special tool,
\href{the GDC Data Transfer Tool}{https://docs.gdc.cancer.gov/Data_Transfer_Tool/Users_Guide/Getting_Started/},
that enables high-performance, potentially parallel, and
resumable downloads. The Data Transfer Tool is an external
program that requires separate download. #' @param gdc_client character(1) name or path to \code{gdc-client}
    executable.  The executable that is used is found through the
    \code{\link{gdc_client}}. See \code{\link{gdc_client}}
    for details on how to set the executable path.

\code{transfer_help()} queries the the command line GDC Data
Transfer Tool, \code{gdc-client}, for available options to be used
in the \code{\link{transfer}} command.
}
\section{Functions}{
\itemize{
\item \code{transfer_help}: 
}}

\examples{
\donttest{
file_manifest = files() \%>\% filter(~ access == "open") \%>\% manifest(size=10)
manifest_file = tempfile()
write.table(file_manifest,file=manifest_file,col.names=TRUE,row.names=FALSE,quote=FALSE)
destination <- transfer(manifest_file)
dir(destination)
# and with authenication
destination <- transfer(manifest_file,token=gdc_token)
}

}
