% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_transfer.R
\name{transfer}
\alias{transfer}
\title{Bulk data download}
\usage{
transfer(manifest, destination_dir = tempfile(), args = character(),
  token = NULL, gdc_client = "gdc-client")
}
\arguments{
\item{manifest}{character(1) file path to manifest created by
\code{manifest()}. See \code{\link{write_manifest}} for 
a simple way to create a manifest file from a data.frame
created with \code{\link{manifest}}.}

\item{destination_dir}{The path into which to place the transfered files.}

\item{args}{character() vector specifying command-line arguments to
be passed to \code{gdc-client}. See \code{\link{transfer_help}} for
possible values. The arguments \code{--manifest}, \code{--dir},
and \code{--token-file} are determined by \code{manifest},
\code{destination_dir}, and \code{token}, respectively, and
should NOT be provided as elements of \code{args}.}

\item{token}{character(1) containing security
token allowing access to restricted data. See
\url{https://gdc-docs.nci.nih.gov/API/Users_Guide/Authentication_and_Authorization/}.
Note that the GDC transfer tool requires a file for data
transfer. Therefore, this token will be written to a temporary
file (with appropriate permissions set).}

\item{gdc_client}{character(1) name or path to \code{gdc-client}
executable. On Windows, use \code{/} or \code{\\\\} as the file
separator.}
}
\value{
character(1) directory path to which the files were
    downloaded.
}
\description{
The GDC maintains a special tool, 
\href{the GDC Data Transfer Tool}{https://docs.gdc.cancer.gov/Data_Transfer_Tool/Users_Guide/Getting_Started/},
that enables high-performance, potentially parallel, and 
resumable downloads. The Data Transfer Tool is an external
program that requires separate download.
}
\examples{
\donttest{
file_manifest = files() \%>\% filter(~ access == "open") \%>\% manifest(size=10)
manifest_file = tempfile()
write.table(file_manifest,file=manifest_file,col.names=TRUE,row.names=FALSE,quote=FALSE)
destination <- transfer(manifest_file)
dir(destination)
# and with authenication
destination <- transfer(manifest_file,token=gdc_token)
}

}

